+++ src/App.js
@@ -1,9 +1,9 @@
 import { Routes, Route } from "react-router-dom";
-import PublicPage from "./routes/PublicPage";
+import PublicPage from "./routes/Pets/Pets";
 
-import Login from "./routes/Login";
+import Login from "./routes/Login/Login";
 import Layout from "./components/Layout";
-import CreatePet from "./routes/CreatePet";
+import CreatePet from "./routes/CreatePet/CreatePet";
 import { AuthProvider, RequireAuth } from "./context/auth-context";
 
 function App() {
