--- src/routes/CreatePet.css
@@ -1,20 +1,0 @@
-.createForm {
-	display: flex;
-	align-items: start;
-	flex-direction: column;
-	max-width: 250px;
-}
-label {
-	padding-bottom: 8px;
-}
-input {
-	width: 250px;
-}
-button {
-	align-self: center;
-}
-.createApp {
-	display: flex;
-
-	justify-content: center;
-}
