+++ src/routes/CreatePet/CreatePet.css
@@ -1,0 +1,32 @@
+.createForm {
+	max-width: 250px;
+	padding-top: 18px;
+	margin-left: auto;
+	margin-right: auto;
+	background-color: #b794f4;
+	padding: 18px;
+	border-radius: 0 0 5px 5px;
+}
+label {
+	padding-bottom: 8px;
+	padding-right: 8px;
+	font-size: 16px;
+	font-family: "Roboto", sans-serif;
+	display: flex;
+	flex-direction: column;
+	align-items: flex-start;
+}
+input {
+	border-radius: 3px;
+	width: 200px;
+}
+button {
+	align-self: center;
+}
+
+select {
+	border-radius: 3px;
+}
+.input-age {
+	width: 30px;
+}
